WIP module finalizer
