+++ test/metadata/test_file_id.py
@@ -7,18 +7,9 @@ from test.conftest import assert_pdf_equal, EPOCH
 HERE = Path(__file__).resolve().parent
 
 
-def test_default_file_id():
-    pdf = FPDF()
-    pdf.creation_date = EPOCH
-    assert (
-        pdf.file_id()
-        == "<A29419ACCA57A4335F99B957670E2C04><A29419ACCA57A4335F99B957670E2C04>"
-    )
-
-
 def test_custom_file_id(tmp_path):
     class PDF(FPDF):
-        def file_id(self):
+        def file_id(self, buffer, creation_date):
             return "<DEADBEEF><DEADBEEF>"
 
     pdf = PDF()
@@ -27,7 +18,7 @@ def test_custom_file_id(tmp_path):
 
 def test_no_file_id(tmp_path):
     class PDF(FPDF):
-        def file_id(self):
+        def file_id(self, buffer, creation_date):
             return None
 
     pdf = PDF()
